web-animations/animation-effect/animation-effect-name.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationEffect name attribute test</title>
<meta name="assert" content="A string used to identify the effect">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationeffect-name">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var effects = [newKeyframeEffect(), newMotionPathEffect()];
    effects.forEach(function(effect) {
        assert_equals(effect.name, '', type(effect) + '.name should be ' +
            '\'\' by defalt');
    });
}, 'Test AnimationEffect.name default value is \'\'');

test(function() {
    var nameValue = 'someName';
    var effects = [newKeyframeEffect({name: nameValue}),
        newMotionPathEffect({name: nameValue})];
    effects.forEach(function(effect) {
        assert_equals(effect.name, nameValue, type(effect) + '.name should be ' +
            'equal to value specified in constructor');
    });
}, 'Test AnimationEffect.name is initialized by construstor');
</script>
</body>
